Create base plotly figure

Create a base plot where traces can later be added.

Usage

base_plot_z_scores(
  dat,
  cog_vars = c("cog_tmta_time", "cog_otmta_time", "cog_otmta_error", "cog_nsf_total",
    "cog_nsf_span", "cog_nsb_total", "cog_nsb_span"),
  fill_values = setNames(RColorBrewer::brewer.pal(n = 7, "RdYlGn"), nm = c("Impaired",
    "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior")),
  shade_descriptions = T,
  fill_alpha = 0.2,
  source = "A"
)

Arguments

dat

data object to get ranges for axes from

cog_vars

vector of variables to infer ranges from without prefix

fill_values

a named vector of length seven with hex color values to use for categories "Impaired", "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior"

shade_descriptions

logical; should ranges for descriptions be shaded?

fill_alpha

opacity value for shaded areas

source

passed to plotly::plot_ly