Get Age Group for Standardized Scores

Get various versions of age groups for finding standardized scores.

Usage

get_age_group(
  age,
  group_type = c("ravlt_trials", "ravlt_recog", "UDS_battery", "otmt", "cowat", "digsym")
)

Arguments

age

numeric vector with ages to be converted to groups

group_type

string indicator which age group to return. Must be one of "ravlt_trials", "ravlt_recog", "UDS_battery", "otmt", "cowat", "digsym"