Various Cognition Measures: means and standard deviations

Averages and standard deviations for a long list of variables. These are by gender, education level, and age group.

Usage

male_female

Format

male_female

variable

Variable name

name

Variable name in WLS REDCap database

UDS_age_group

Age groups. 1: <60, 2: 60-69, 3: 70-79, 4: 80-89, 5: >90

edu_group

Education group by years of education. A: <12, B: 13-15, C: 16, D: >16

cog_sex

Sex group

m

Mean of name for age group UDS_age_group, eduction group education, and gender cog_sex

sd

Standard deviation of name for age group UDS_age_group, eduction group education, and gender cog_sex