Prepare data from REDCap

This function prepares data downloaded from REDCap to be presented using main_table(). It removes records without a valid age before running add_standardized_scores(), which adds age groups and standardized scores.

Usage

prep_data(
  decrypt_password,
  file,
  nas = c("NA", "N/A", "-", " ", "", "#REF!", "17/22", "#DIV/0!", "#VALUE!", "0,5",
    "23a", "28b")
)

Arguments

decrypt_password

password to decrypt REDCap API Token so data can be fetched straight from REDCap.

file

path to file downloaded from REDCap

nas

character vector of strings that should be treated as NA

Examples

if (FALSE) { # \dontrun{
prep_data(
  file = "~/Downloads/WisconsinLongitudina_DATA_2024-08-10_1500.csv",
  nas = c("NA", "N/A", "-", " ", "", "#REF!", "17/22", "#DIV/0!", "#VALUE!")
)
} # }