Create Row Labels for Main Table

The main table of interest uses some of the observed data in the row labels. This function creates the row labels as a named vector for easy use in stringr::str_replace_all to create a new column with the desired row labels. All parameters should be "raw" values.

Usage

var_labels(
  cog_cdr_sob = NA,
  cog_tmta_err = NA,
  cog_tmta_lines = NA,
  cog_benson_retained = NA,
  cog_benson_recog = NA,
  cog_ravlt_a1 = NA,
  cog_ravlt_a2 = NA,
  cog_ravlt_a3 = NA,
  cog_ravlt_a4 = NA,
  cog_ravlt_a5 = NA,
  cog_ravlt_recog_tp = NA,
  cog_ravlt_recog_tn = NA,
  cog_tmtb_err = NA,
  cog_tmtb_lines = NA,
  cog_craft_verb_retain = NA,
  cog_craft_par_retain = NA
)

Arguments

cog_cdr_sob

Raw observed value of CDR Sob

cog_tmta_err

number of errors made on Trail Making Test A

cog_tmta_lines

lines completed on Trail Making Test A

cog_benson_retained

percent retained on Benson

cog_benson_recog

Benson recognized

cog_ravlt_a1

Rey AVLT Trial 1

cog_ravlt_a2

Rey AVLT Trial 2

cog_ravlt_a3

Rey AVLT Trial 3

cog_ravlt_a4

Rey AVLT Trial 4

cog_ravlt_a5

Rey AVLT Trial 5

cog_ravlt_recog_tp

Rey AVLT Recognition True Positives

cog_ravlt_recog_tn

Rey AVLT Recognition True Negatives

cog_tmtb_err

number of errors made on Trail Making Test B

cog_tmtb_lines

lines completed on Trail Making Test B